Changed: Extract class `BufferStyler`
Extract `setOrClearEffect()` to class. `BufferStyler` class satisfy Single responsibility principle.
